figgy_backend/requirements.txt
Flask==3.0.3
pymongo==4.5.0
flask-cors==4.0.0
python-dotenv==1.0.0
dnspython==2.4.2
razorpay==1.4.1
Flask-APScheduler==1.13.1
requests==2.31.0
Werkzeug==3.0.3
firebase-admin==6.2.0
apscheduler==3.10.1